style: страница курса перенесена в другое место
