+++ src/app/pages/course-page/components/course-header/course-header.component.html
