+++ src/app/pages/course-page/components/course-header/course-header.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
 
 import { EFullRoutes } from 'src/app/shared/router-paths';
-import { TCourse } from 'src/app/pages/courses-page/pages/course-page/course-page.interface';
+import { TCourse } from 'src/app/pages/course-page/course-page.interface';
 
 @Component({
   selector: 'app-course-header',
