+++ src/app/pages/course-page/components/course-header/course-header.module.ts
@@ -4,11 +4,11 @@ import { CdkListbox } from '@angular/cdk/listbox';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 import { RouterLink } from '@angular/router';
 
-import { CoursePracticeItemComponent } from 'src/app/pages/courses-page/pages/course-page/components/course-practice-item/course-practice-item.component';
+import { CourseHeaderComponent } from 'src/app/pages/course-page/components/course-header/course-header.component';
 
 @NgModule({
-  declarations: [CoursePracticeItemComponent],
+  declarations: [CourseHeaderComponent],
   imports: [NgStyle, CdkListbox, NzIconModule, NgIf, RouterLink],
-  exports: [CoursePracticeItemComponent],
+  exports: [CourseHeaderComponent],
 })
-export class CoursePracticeItemModule {}
+export class CourseHeaderModule {}
