+++ src/app/pages/course-page/components/course-practice-item/course-practice-item.component.scss
