+++ src/app/pages/course-page/course-page.component.html
