+++ src/app/pages/course-page/course-page.component.scss
