+++ src/app/pages/course-page/course-page.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { ActivatedRoute } from '@angular/router';
 
-import { CoursePageService } from 'src/app/pages/courses-page/pages/course-page/course-page.service';
+import { CoursePageService } from 'src/app/pages/course-page/course-page.service';
 import { SideBarService } from 'src/app/components/side-bar/side-bar.service';
 import { ERoles } from 'src/app/core/role/role.enum';
 import { EFullRoutes, ERoutesIds } from 'src/app/shared/router-paths';
