+++ src/app/pages/course-page/course-page.dto.ts
