+++ src/app/pages/course-page/course-page.interface.ts
