+++ src/app/pages/courses-page/components/course-item/courses-page-item.component.ts
@@ -2,7 +2,7 @@ import { Component, Input } from '@angular/core';
 
 import { CoursesPageItemService } from 'src/app/pages/courses-page/components/course-item/courses-page-item.service';
 import { EFullRoutes } from 'src/app/shared/router-paths';
-import { TCourse } from 'src/app/pages/courses-page/pages/course-page/course-page.interface';
+import { TCourse } from 'src/app/pages/course-page/course-page.interface';
 
 @Component({
   selector: 'app-courses-page-item',
