+++ src/app/pages/practice-page/practice-page.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 
-import { TCourse } from 'src/app/pages/courses-page/pages/course-page/course-page.interface';
+import { TCourse } from 'src/app/pages/course-page/course-page.interface';
 import { mockCourses } from 'src/app/pages/courses-page/mock-courses';
-import { TCourseDto } from 'src/app/pages/courses-page/pages/course-page/course-page.dto';
+import { TCourseDto } from 'src/app/pages/course-page/course-page.dto';
 import { TPractice } from 'src/app/pages/practice-page/practice-page.interface';
 
 import { BehaviorSubject, Observable } from 'rxjs';
